Function added to detect date change

+++ assets/js/newDayupdate.js
@@ -1,50 +1,61 @@
-const cron = require('node-cron');
 const Habits = require('../../models/habits');
 const Status = require('../../models/status');
-
-const monthNames = [
-    "January", "February", "March", "April", "May", "June",
-    "July", "August", "September", "October", "November", "December"
-    ];
 
 (async () => {
   try {
     let currentDate = new Date();
-    const month = monthNames[currentDate.getMonth()];
-    const day = currentDate.getDate();
-    const formattedDate = `${month} ${day}`;
+    let previousDate = currentDate;
 
-    let dateToRemove = new Date();
-    dateToRemove.setDate(dateToRemove.getDate() - 7);
-    const RemoveMonth = monthNames[dateToRemove.getMonth()];
-    const RemoveDay = dateToRemove.getDate();
-    const RemoveDate = `${RemoveMonth} ${RemoveDay}`;
+    while (true) {
+      currentDate = new Date();
 
-    let habits = await Habits.find({});
+      if (currentDate.getDate() !== previousDate.getDate()) {
 
-    for (let habit of habits) {
-      let status = await Status.findOne({ date: RemoveDate, habit: habit._id });
+        console.log('function run')
+        const month = monthNames[currentDate.getMonth()];
+        const day = currentDate.getDate();
+        const formattedDate = `${month} ${day}`;
 
-      if (status) {
-        let dateID = status._id;
+        const dateToRemove = new Date(previousDate);
+        dateToRemove.setDate(dateToRemove.getDate() - 7);
+        const RemoveMonth = monthNames[dateToRemove.getMonth()];
+        const RemoveDay = dateToRemove.getDate();
+        const RemoveDate = `${RemoveMonth} ${RemoveDay}`;
 
-        await Status.deleteOne({ _id: dateID });
+        let habits = await Habits.find({});
 
-        await Habits.findByIdAndUpdate(habit._id, { $pull: { status: dateID } });
+        for (let habit of habits) {
+          let status = await Status.findOne({ date: RemoveDate, habit: habit._id });
+
+          if (status) {
+            let dateID = status._id;
+
+            await Status.deleteOne({ _id: dateID });
+
+            await Habits.findByIdAndUpdate(habit._id, { $pull: { status: dateID } });
+          }
+
+          let newStatus = await Status.create({
+            date: formattedDate,
+            datestatus: 'Not Started',
+            habit: habit._id
+          });
+
+          habit.status.push(newStatus._id);
+          await habit.save();
+        }
       }
 
-      let newStatus = await Status.create({
-        date: formattedDate,
-        datestatus: 'Not Started',
-        habit: habit._id
-      });
+      previousDate = currentDate;
 
-      habit.status.push(newStatus._id);
-      await habit.save();
+      // Delay for 1 minute (adjust as needed)
+      await delay(60000);
     }
   } catch (error) {
     console.log('Error', error);
   }
 })();
 
-
+function delay(ms) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
